test(core): consolidate DOM timer regressions
