--- e2e/dom/fixtures/test/nonIsolatedStaleTimerA.test.ts
@@ -1,4 +1,0 @@
-import { it } from '@rstest/core';
-import { runStaleTimerPhase } from './nonIsolatedStaleTimer';
-
-it('retains a timer wrapper from the first environment', runStaleTimerPhase);
