--- e2e/dom/fixtures/test/nonIsolatedStaleTimerB.test.ts
@@ -1,4 +1,0 @@
-import { it } from '@rstest/core';
-import { runStaleTimerPhase } from './nonIsolatedStaleTimer';
-
-it('reports errors from the retained timer wrapper', runStaleTimerPhase);
