+++ e2e/dom/fixtures/test/nonIsolatedTimers.ts
@@ -1,22 +1,44 @@
-import { promisify } from 'node:util';
 import { expect, rstest } from '@rstest/core';
 
 const phaseKey = Symbol.for('rstest.jsdom.timer-phase');
 const staleTimerKey = Symbol.for('rstest.jsdom.stale-timer');
 const completedTimerKey = Symbol.for('rstest.jsdom.completed-timer');
-const userAbortKey = Symbol.for('rstest.jsdom.user-abort');
-const noOpClearKey = Symbol.for('rstest.jsdom.no-op-clear');
+const staleWrapperKey = Symbol.for('rstest.jsdom.stale-timer-wrapper');
 
 export const runTimerPhase = async () => {
   const phase = (Reflect.get(process, phaseKey) as number | undefined) ?? 0;
+  const reportsStaleError = process.env.RSTEST_STALE_TIMER_ERROR === '1';
 
   if (phase === 0) {
+    if (reportsStaleError) {
+      Reflect.set(process, staleWrapperKey, setTimeout);
+      Reflect.set(process, phaseKey, 1);
+      return;
+    }
     rstest.useFakeTimers({ now: 0 });
     Reflect.set(process, phaseKey, 1);
     return;
   }
 
   if (phase === 1) {
+    if (reportsStaleError) {
+      const staleSetTimeout = Reflect.get(
+        process,
+        staleWrapperKey,
+      ) as typeof setTimeout;
+      Reflect.deleteProperty(process, phaseKey);
+      Reflect.deleteProperty(process, staleWrapperKey);
+
+      const timeout = staleSetTimeout(() => {}, 60_000);
+      expect(timeout.hasRef()).toBe(false);
+      clearTimeout(timeout);
+      staleSetTimeout(() => {
+        throw new Error('retained stale timer error');
+      }, 0);
+      await new Promise((resolve) => setTimeout(resolve, 20));
+      return;
+    }
+
     rstest.useFakeTimers({ now: 0 });
     rstest.useRealTimers();
     Reflect.set(process, staleTimerKey, false);
@@ -25,16 +47,6 @@ export const runTimerPhase = async () => {
       const timer = setTimeout(() => resolve(timer), 0);
     });
     Reflect.set(process, completedTimerKey, completedTimer);
-    const controller = new AbortController();
-    const userReason = new Error('non-isolated user abort');
-    const userAbort = promisify(setTimeout)(60_000, undefined, {
-      signal: controller.signal,
-    }).catch((error) => error);
-    controller.abort(userReason);
-    Reflect.set(process, userAbortKey, userAbort);
-    const noOpClear = promisify(setTimeout)(50, 'resolved');
-    Reflect.apply(clearTimeout, globalThis, [noOpClear]);
-    Reflect.set(process, noOpClearKey, noOpClear);
     Reflect.set(process, phaseKey, 2);
     return;
   }
@@ -47,31 +59,9 @@ export const runTimerPhase = async () => {
   expect(completedTimer.hasRef()).toBe(false);
   clearTimeout(completedTimer);
 
-  const userAbort = Reflect.get(process, userAbortKey) as Promise<unknown>;
-  await expect(
-    Promise.race([
-      userAbort,
-      new Promise((resolve) => setTimeout(() => resolve('pending'), 100)),
-    ]),
-  ).resolves.toMatchObject({
-    cause: expect.any(Error),
-    code: 'ABORT_ERR',
-    name: 'AbortError',
-  });
-
-  const noOpClear = Reflect.get(process, noOpClearKey) as Promise<unknown>;
-  await expect(
-    Promise.race([
-      noOpClear,
-      new Promise((resolve) => setTimeout(() => resolve('pending'), 100)),
-    ]),
-  ).resolves.toBe('pending');
-
   await new Promise((resolve) => setTimeout(resolve, 100));
   expect(Reflect.get(process, staleTimerKey)).toBe(false);
   Reflect.deleteProperty(process, phaseKey);
   Reflect.deleteProperty(process, staleTimerKey);
   Reflect.deleteProperty(process, completedTimerKey);
-  Reflect.deleteProperty(process, userAbortKey);
-  Reflect.deleteProperty(process, noOpClearKey);
 };
