+++ e2e/dom/fixtures/test/timers.test.ts
@@ -26,46 +26,9 @@ it('keeps Node timer handles while tracking real timers', async () => {
   expect(receiver.unref).toBeTypeOf('function');
 });
 
-it('clears numeric Node timer ids without running their callbacks', async () => {
-  let timeoutCalled = false;
-  const timeout = setTimeout(() => {
-    timeoutCalled = true;
-  }, 10);
-  clearInterval(Number(timeout));
-
-  let intervalCalls = 0;
-  const interval = setInterval(() => {
-    intervalCalls++;
-  }, 1);
-  clearTimeout(Number(interval));
-
-  await new Promise((resolve) => nodeSetTimeout(resolve, 20));
-  expect(timeoutCalled).toBe(false);
-  expect(intervalCalls).toBe(0);
-});
-
-it('preserves Node setTimeout utility behavior', async () => {
+it('preserves promisified timeout behavior', async () => {
   const sleep = promisify(setTimeout);
   await expect(sleep(1, 'done')).resolves.toBe('done');
-  await expect(sleep(1, 'unref', { ref: false })).resolves.toBe('unref');
-  const inheritedOptions = Object.create(null);
-  Object.defineProperty(inheritedOptions, 'ref', { value: false });
-  await expect(sleep(1, 'inherited ref', inheritedOptions)).resolves.toBe(
-    'inherited ref',
-  );
-  await expect(
-    Reflect.apply(sleep, undefined, [1, undefined, { signal: null }]),
-  ).rejects.toMatchObject({ code: 'ERR_INVALID_ARG_TYPE' });
-
-  const NativeAbortController = AbortController;
-  try {
-    Reflect.set(globalThis, 'AbortController', undefined);
-    await expect(sleep(1, 'captured controller')).resolves.toBe(
-      'captured controller',
-    );
-  } finally {
-    Reflect.set(globalThis, 'AbortController', NativeAbortController);
-  }
 
   const addEventListener = AbortSignal.prototype.addEventListener;
   try {
@@ -89,37 +52,11 @@ it('preserves Node setTimeout utility behavior', async () => {
     code: 'ABORT_ERR',
     cause: 'test abort',
   });
-
-  let errorCode: string | undefined;
-  try {
-    Reflect.apply(setTimeout, globalThis, ['invalid', 0]);
-  } catch (error) {
-    errorCode = (error as { code?: string }).code;
-  }
-  expect(errorCode).toBe('ERR_INVALID_ARG_TYPE');
 });
 
 it('cleans up pending promisified timeouts during teardown', () => {
   const sleep = promisify(setTimeout);
   void sleep(60_000).then(() => {});
-});
-
-it('supports nested timers and cross-clearing', async () => {
-  let intervalCalls = 0;
-  const interval = setInterval(() => {
-    intervalCalls++;
-  }, 0);
-  clearTimeout(interval);
-  await new Promise((resolve) => nodeSetTimeout(resolve, 20));
-  expect(intervalCalls).toBe(0);
-
-  await new Promise<void>((resolve) => {
-    const outer = setTimeout(() => {
-      const inner = setTimeout(resolve, 0);
-      expect(typeof inner).toBe('object');
-    }, 0);
-    expect(typeof outer).toBe('object');
-  });
 });
 
 it('keeps DOM-owned animation frame timers working', async () => {
