--- e2e/dom/unhandledTimerError.test.ts
@@ -1,14 +1,0 @@
-import { describe, it } from '@rstest/core';
-import { runCli } from './utils';
-
-describe('jsdom timer errors', () => {
-  it('fails for observed errors that are not prevented', async () => {
-    const { expectExecFailed, expectStderrLog } = await runCli(
-      'test/unhandledTimerError',
-      'jsdom',
-    );
-
-    await expectExecFailed();
-    expectStderrLog('Timer callback threw undefined');
-  });
-});
